% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_authentication.R
\name{addToGitignore}
\alias{addToGitignore}
\title{Add to .gitignore}
\usage{
addToGitignore(addGitignore = NULL)
}
\arguments{
\item{addGitignore}{Option to add the credentials to the .gitignore file (Optional)}
}
\value{
Adds credentials to the .gitignore file.
}
\description{
Adds saved credentials to .gitignore so they are not saved.
}
\details{
By default, this does not add to or create a .gitignore file.
}
