% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_authentication.R
\name{apiDetails}
\alias{apiDetails}
\title{Google API Authentication Details}
\usage{
apiDetails()
}
\value{
lists all credentials needed for loading Adwords data
}
\description{
Enter the currently available Google Login Details
}
