% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_xml_request.R
\name{buildXmlHeader}
\alias{buildXmlHeader}
\title{Build XML Header}
\usage{
buildXmlHeader(myMcc, userAgent, developerToken, apiVersion)
}
\arguments{
\item{myMcc}{The Manager Account (MCC) needed to access the Adwords API}

\item{userAgent}{The user agent required for the header of the XML request}

\item{developerToken}{The developerToken required for accessing the API. Found in the API settings of the Adwords Account.}

\item{apiVersion}{Version of the API being used.}
}
\value{
Returns the XML Header for the Envelope
}
\description{
Build the XML Header for the XML Envelope.
}
\examples{
buildXmlHeader("123-456-7890", "myUserAgent", "myD3v3l0p3r70k3n", "v201806")
}
