% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_authentication.R
\name{generateAdwordsToken}
\alias{generateAdwordsToken}
\title{Initial Adwords Token}
\usage{
generateAdwordsToken(saveNewToken = NULL, addGitignore = NULL)
}
\arguments{
\item{saveNewToken}{Option to save the credentials to the file. (Optional)}

\item{addGitignore}{Option to add the credentials to the .gitignore file (Optional)}
}
\value{
credentials entered and generates the initial token
}
\description{
Generates the initial Adwords Token
}
