% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{getReportData}
\alias{getReportData}
\title{Retrieve Adwords Data}
\usage{
getReportData(reportType, startDate, endDate, clientCustomerId, credentials,
  attributes = NULL, segment = NULL, metrics = NULL, where = NULL,
  apiVersion = NULL, useRawEnumValues = NULL,
  includeZeroImpressions = NULL, useRequestedHeaders = NULL)
}
\arguments{
\item{reportType}{The report being queries.}

\item{startDate}{The start date for the request.}

\item{endDate}{The end date for the request.}

\item{clientCustomerId}{The Client Customer Id to retrieve data for. This should be in the format ("123-456-7890").}

\item{credentials}{The credentials for accessing the Adwords API.}

\item{attributes}{The attributes of the report.}

\item{segment}{The segment of the report.}

\item{metrics}{The metrics of the report.}

\item{where}{The conditions of the report. (Optional)}

\item{apiVersion}{The API version that will be accessed. Defaults to "v201806".}

\item{useRawEnumValues}{Use specified column headers used with the request, or the Adwords Display Names for column headers. Defaults to TRUE.}

\item{includeZeroImpressions}{Whether Entries with zero impressions should be included in the data. Defaults to FALSE.}

\item{useRequestedHeaders}{Use the headers used to retrieve the Adwords data or the Display names.}
}
\value{
Returns Data for the requested report.
}
\description{
Retrieve the Adwords Data from the Adwords API
}
