% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_xml_request.R
\name{getXmlRequest}
\alias{getXmlRequest}
\title{Send XML Request}
\usage{
getXmlRequest(adwordsService, xmlEnvelope, credentials, saveNewToken = NULL,
  addGitignore = NULL, apiVersion = NULL)
}
\arguments{
\item{adwordsService}{The Adwords service that is being requested}

\item{xmlEnvelope}{The full XML request to be sent to the Adwords API}

\item{credentials}{The list that contains the access token. This does not need to be valid.}

\item{saveNewToken}{Option to save the new access token if access token has been refreshed. (Optional)}

\item{addGitignore}{Option to add new access token to Gitignore if required. (Optional)}

\item{apiVersion}{The version of the API being requested. Defaults to v201806.}
}
\value{
Returns the XML output from the API request.
}
\description{
Send XML Get Request to get a XML response.
}
\details{
Please note that sending your XML request to the API does check and refresh the token automatically, if necessary.
}
