% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_authentication.R
\name{loadAdwordsToken}
\alias{loadAdwordsToken}
\title{Load Adwords Token}
\usage{
loadAdwordsToken(saveNewToken = NULL, addGitignore = NULL)
}
\arguments{
\item{saveNewToken}{Option to save the credentials to the file. (Optional)}

\item{addGitignore}{Option to add the credentials to the .gitignore file (Optional)}
}
\value{
loads credentials from existing token or new token
}
\description{
Loads a pre-existing token if it exists, or generates a token if it does not exist.
}
