% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managed_customer_service.R
\name{ManagedCustomerService_serviceSelector}
\alias{ManagedCustomerService_serviceSelector}
\title{Customer Manager Service Selector}
\usage{
ManagedCustomerService_serviceSelector(ManagedCustomerService_Selector_fields,
  apiVersion)
}
\arguments{
\item{ManagedCustomerService_Selector_fields}{The fields required from the ManagedCustomerService API}

\item{apiVersion}{Version of the API for the request. Argument is passed through from buildXmlEnvelope}
}
\value{
Returns the XML text for the Service Selector
}
\description{
ManagedCustomerService.serviceSelector
}
\details{
This is used to specify the type of information to return.
}
\examples{
ManagedCustomerService_get(c("Name", "CustomerId"), "v201806")
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/ManagedCustomerService.Selector}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
