% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_authentication.R
\name{refreshAdwordsToken}
\alias{refreshAdwordsToken}
\title{Refresh Adwords Token}
\usage{
refreshAdwordsToken(credentials, saveNewToken = NULL, addGitignore = NULL)
}
\arguments{
\item{credentials}{The current available token credentials which includes the access token}

\item{saveNewToken}{Option to save the credentials to the file. (Optional)}

\item{addGitignore}{Option to add the credentials to the .gitignore file (Optional)}
}
\value{
The credentials containing the access token that needs updating using the refresh token
}
\description{
Refresh the Adwords token to access the API.
}
