% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_categoryProductsAndServicesSearchParameter}
\alias{TargetingIdeaService_categoryProductsAndServicesSearchParameter}
\title{TargetingIdeaService - Category Products and Services Search Parameter}
\usage{
TargetingIdeaService_categoryProductsAndServicesSearchParameter(
categoryProductsAndServices)
}
\arguments{
\item{categoryProductsAndServices}{A keyword category ID (Integer) in the "Products and Services" taxonomy that all search results should belong to.}
}
\value{
Returns the XML text for CategoryProductsAndServicesSearchParameter Search Parameter
}
\description{
TargetingIdeaService.CategoryProductsAndServicesSearchParameter
}
\details{
Idea Type supported: KEYWORD

Request Type supported: IDEAS
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.CategoryProductsAndServicesSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
