% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_competitionSearchParameter}
\alias{TargetingIdeaService_competitionSearchParameter}
\title{TargetingIdeaService - Competition Search Parameter}
\usage{
TargetingIdeaService_competitionSearchParameter(competition)
}
\arguments{
\item{competition}{Vector of Levels of competition that should be included in the results}
}
\value{
This is not currently supported.
}
\description{
TargetingIdeaService.CompetitionSearchParameter
}
\details{
LOW - competition rate [0.0000, 0.3333]

MEDIUM - competition rate (0.3333, 0.6667]

HIGH - competition rate (0.6667, 1.0000]
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.CompetitionSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
