% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_ideaTextFilterSearchParameter}
\alias{TargetingIdeaService_ideaTextFilterSearchParameter}
\title{TargetingIdeaService - Indea Text Filter Search Parameter}
\usage{
TargetingIdeaService_ideaTextFilterSearchParameter(ideaTextFilter_included,
  ideaTextFilter_excluded)
}
\arguments{
\item{ideaTextFilter_included}{The vector of strings that should be included}

\item{ideaTextFilter_excluded}{The vector of strings that should be excluded}
}
\value{
Returns the XML text for IdeaTextFilterSearchParameter Search Parameter
}
\description{
TargetingIdeaService.IdeaTextFilterSearchParameter
}
\details{
A maximum 200 included/excluded strings can be used
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.IdeaTextFilterSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
