% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_ideaType}
\alias{TargetingIdeaService_ideaType}
\title{TargetingIdeaService - Idea Types}
\usage{
TargetingIdeaService_ideaType()
}
\value{
Returns the XML text for Idea Type. By default, the Idea Type will be "KEYWORD" as this is the only option.
}
\description{
TargetingIdeaService.IdeaType
}
\details{
Idea Type is currently limited to "KEYWORD" only.
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.IdeaType}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
