% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_includeAdultContentSearchParameter}
\alias{TargetingIdeaService_includeAdultContentSearchParameter}
\title{TargetingIdeaService - Include Adult Content Search Parameter}
\usage{
TargetingIdeaService_includeAdultContentSearchParameter()
}
\value{
Returns the XML text for IncludeAdultContentSearchParameter Search Parameter if Adult content is requested to be included.
}
\description{
TargetingIdeaService.IncludeAdultContentSearchParameter
}
\details{
If used, this function specifies whether adult content should be returned.
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.IncludeAdultContentSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
