% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_languageSearchParameter}
\alias{TargetingIdeaService_languageSearchParameter}
\title{TargetingIdeaService - Language Search Parameter}
\usage{
TargetingIdeaService_languageSearchParameter(language, apiVersion)
}
\arguments{
\item{language}{The Language (or Criterion ID) of the desired languages being targeted in the results.}

\item{apiVersion}{The API version.}
}
\value{
Returns the XML text for LanguageSearchParameter Search Parameter
}
\description{
TargetingIdeaService.LanguageSearchParameter
}
\details{
Multiple languages can be chosen, as a list. This must be in vector format.
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.LanguageSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
