% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_locationSearchParameter}
\alias{TargetingIdeaService_locationSearchParameter}
\title{TargetingIdeaService - Location Search Parameter}
\usage{
TargetingIdeaService_locationSearchParameter(location, apiVersion)
}
\arguments{
\item{location}{The location (or Criterion ID) of the desired locations being targeted in the results.}

\item{apiVersion}{The API version.}
}
\value{
Returns the XML text for LocationSearchParameter Search Parameter
}
\description{
TargetingIdeaService.LocationSearchParameter
}
\details{
Multiple locations can be chosen.
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.LocationSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
