% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_networkSearchParameter}
\alias{TargetingIdeaService_networkSearchParameter}
\title{TargetingIdeaService - Network Search Parameter}
\usage{
TargetingIdeaService_networkSearchParameter(network_GoogleSearch,
  network_SearchNetwork, apiVersion)
}
\arguments{
\item{network_GoogleSearch}{Whether results should be targeting Google Search}

\item{network_SearchNetwork}{Whether results should be targeting the Google Search Network (AFS)}

\item{apiVersion}{The API version}
}
\value{
Returns the XML text for NetworkSearchParameter Search Parameter
}
\description{
TargetingIdeaService.NetworkSearchParameter
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.NetworkSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
