% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_paging}
\alias{TargetingIdeaService_paging}
\title{TargetingIdeaService - Paging Request}
\usage{
TargetingIdeaService_paging(pagingStartIndex, pagingNumberResults, apiVersion)
}
\arguments{
\item{pagingStartIndex}{Index of the first result to return. (Optional)}

\item{pagingNumberResults}{Maximum number of results to return. This figure is limited to 700. (Optional)}

\item{apiVersion}{Version of the API for the request}
}
\value{
Returns the XML text for paging
}
\description{
TargetingIdeaService.Paging
}
\details{
\emph{pagingNumberResults} is limited to 700. Any figure over 700 will be reduced to 700.
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.Paging}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
