% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_relatedToQuerySearchParameter}
\alias{TargetingIdeaService_relatedToQuerySearchParameter}
\title{TargetingIdeaService - Related to Query Search Parameter}
\usage{
TargetingIdeaService_relatedToQuerySearchParameter(relatedToQuery)
}
\arguments{
\item{relatedToQuery}{The vector of queries to post to the API as a Search Parameter.}
}
\value{
Returns the XML text for RelatedToQuerySearchParameter Search Parameter
}
\description{
TargetingIdeaService.RelatedToQuerySearchParameter
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.RelatedToQuerySearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
