% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_relatedToUrlSearchParameter}
\alias{TargetingIdeaService_relatedToUrlSearchParameter}
\title{TargetingIdeaService - Related to URL Search Parameter}
\usage{
TargetingIdeaService_relatedToUrlSearchParameter(relatedToUrl)
}
\arguments{
\item{relatedToUrl}{The vector of URLs to post to the API, which results will be related to.}
}
\value{
Returns the XML text for RelatedToUrlSearchParameter Search Parameter
}
\description{
TargetingIdeaService.RelatedToUrlSearchParameter
}
\details{
For KEYWORD queries, only one URL may be submitted.
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.RelatedToUrlSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
