% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_requestedAttributeTypes}
\alias{TargetingIdeaService_requestedAttributeTypes}
\title{TargetingIdeaService - Attribute Types}
\usage{
TargetingIdeaService_requestedAttributeTypes(attributeTypes)
}
\arguments{
\item{attributeTypes}{Vector of Attributes Types.}
}
\value{
Returns the XML text for attributeTypes
}
\description{
TargetingIdeaService.AttributeType
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.AttributeType}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
