% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_requestType}
\alias{TargetingIdeaService_requestType}
\title{TargetingIdeaService - Request Types}
\usage{
TargetingIdeaService_requestType(requestType)
}
\arguments{
\item{requestType}{Represents the type of the request.}
}
\value{
Returns the XML text for request type.
}
\description{
TargetingIdeaService.RequestType
}
\details{
The \emph{Request Type} is limited to "IDEAS" and "STATS".

"IDEAS" is currently unsupported.
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.RequestType}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
