% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_searchVolumeSearchParameter}
\alias{TargetingIdeaService_searchVolumeSearchParameter}
\title{TargetingIdeaService - Search Volume Search Parameter}
\usage{
TargetingIdeaService_searchVolumeSearchParameter(searchVolumeMinimum,
  searchVolumeMaximum)
}
\arguments{
\item{searchVolumeMinimum}{The minimum Search Volume that should be targeted.}

\item{searchVolumeMaximum}{The maximum Search Volume that should be targeted.}
}
\value{
Returns the XML text for SearchVolumeSearchParameter Search Parameter
}
\description{
TargetingIdeaService.SearchVolumeSearchParameter
}
\details{
Using both parameters returns results with search volume in this regon.
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.SearchVolumeSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
