% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_idea_service.R
\name{TargetingIdeaService_seedAdGroupIdSearchParameter}
\alias{TargetingIdeaService_seedAdGroupIdSearchParameter}
\title{TargetingIdeaService - Seed Ad Group Id Search Parameter}
\usage{
TargetingIdeaService_seedAdGroupIdSearchParameter(seedAdGroupId)
}
\arguments{
\item{seedAdGroupId}{The Adgroup ID that should be used as a seed for generating new ideas.}
}
\value{
Returns the XML text for SeedAdGroupIdSearchParameter Search Parameter
}
\description{
TargetingIdeaService.SeedAdGroupIdSearchParameter
}
\references{
Documentation content adapted from \href{https://developers.google.com/adwords/api/docs/reference/v201802/TargetingIdeaService.SeedAdGroupIdSearchParameter}{Google Adwords API documentation} under license from \href{https://creativecommons.org/licenses/by/3.0/}{CC BY 3.0}
}
